+++ Servers/GameServer/GameEventFactory.cpp
@@ -17,7 +17,7 @@ using namespace SEGSEvents;
 
 GameLinkEvent *GameEventFactory::EventFromStream(BitStream &bs)
 {
-    size_t read_pos = bs.GetReadPos();
+    uint32_t read_pos = bs.GetReadPos();
     //size_t bits_avail = bs.GetReadableBits();
     GameLinkEvent *ev = CRUD_EventFactory::EventFromStream(bs);
     if(ev) // base class created the event
